requirements.txt
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
transformers==4.51.1
# vllm==0.8.4
wandb
packaging>=20.0
uvicorn
fastapi
qwen-vl-utils[decord]
supabase>=2.0.0
python-dotenv>=1.0.0